Moved some code from operations.rb to lib
