+++ lib/kaigara.rb
@@ -10,3 +10,4 @@ require 'kaigara/application'
 require 'controllers/sysops'
 require 'kaigara/client'
 require 'kaigara/metadata'
+require 'kaigara/shell'
