+++ lib/kaigara/shell.rb
@@ -1,0 +1,7 @@
+module Kaigara
+  class Shell
+    include Thor::Base
+    include Thor::Actions
+    include Thor::Shell
+  end
+end
